pyproject.toml
[project]
name = "production-issue-investigator"
version = "0.1.0"
description = "AI agent for investigating production issues using Claude Agent SDK"
requires-python = ">=3.10"
dependencies = [
    "anthropic>=0.40.0",
    "claude-agent-sdk>=0.1.35",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "PyGithub>=2.1.1",
    "python-dateutil>=2.8.2",
    "pytz>=2023.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agents", "utils"]
